🐞 fix(padding): left padding side
Padding on right side is wrong according to https://huggingface.co/docs/transformers/llm_tutorial#wrong-padding-side.
When model inferences, it generates from left to right. If we use batch decoding, we also need to pad inputs. Therefore, we need to padding on left to keep consistency between training and testing.